Add method to separate fnirs data into segments.
The trials are tagged with events that are generated by user input. For
this particular study the data should be tagged with three events. There
is a period of ~20s of quiet stance, ~80s walking, and then more quiet
stance at the end. The period of walking is also split in half to create
an 'early' and 'late' phase.