Remove code that subtracts signal mean from entire signal.
I was taking the mean of the signal and then subtracting this from the
entire signal, as this is how you remove the DC offset (effectively removing
the 0hz component of a signal). However, I think with the bandpass filter
that I am using this is not necessary, so I removed it.