fix: Switch the order of processing steps.
I changed the order of the processing steps to have the TDDR
algorithm be performed on both the short and the long data as a first
step. This makes sense since MAs are present in both the short and long
data and it would be sensible to remove those before proceeding with
the subsequent processing steps. Before I did ssc_regression first, then
TDDR only on the long data.